+++ course_discovery/apps/course_metadata/tests/test_models.py
@@ -18,7 +18,8 @@ from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.core.management import call_command
 from django.db import IntegrityError, transaction
-from django.test import TestCase, override_settings
+from django.test import TestCase, TransactionTestCase, override_settings
+from django_celery_results.models import TaskResult
 from edx_django_utils.cache import RequestCache
 from edx_toggles.toggles.testutils import override_waffle_switch
 from freezegun import freeze_time
@@ -33,7 +34,7 @@ from course_discovery.apps.core.models import Currency
 from course_discovery.apps.core.tests.helpers import make_image_file
 from course_discovery.apps.core.utils import SearchQuerySetWrapper
 from course_discovery.apps.course_metadata.choices import (
-    CourseRunRestrictionType, CourseRunStatus, ExternalProductStatus, ProgramStatus
+    BulkOperationType, CourseRunRestrictionType, CourseRunStatus, ExternalProductStatus, ProgramStatus
 )
 from course_discovery.apps.course_metadata.constants import SUBDIRECTORY_PROGRAM_SLUG_FORMAT_REGEX as slug_format
 from course_discovery.apps.course_metadata.models import (
@@ -2015,91 +2016,6 @@ class OrganizationTests(TestCase):
         org.description = 'test description'
         org.save()
         assert data_modified_timestamp < org.data_modified_timestamp
-
-    def test_org_enterprise_subscription_inclusion_toggle_course(self):
-        """Test that toggling an org's enterprise_subscription_inclusion value will turn courses in the org on"""
-        org = factories.OrganizationFactory(enterprise_subscription_inclusion=True)
-        course_type = CourseType.objects.filter(slug=CourseType.VERIFIED_AUDIT).first()
-        course = factories.CourseFactory(enterprise_subscription_inclusion=True, type=course_type)
-        course.authoring_organizations.add(org)
-        course.save()
-
-        org.enterprise_subscription_inclusion = False
-        org.save()
-
-        course.refresh_from_db()
-        assert course.enterprise_subscription_inclusion is False
-
-    def test_org_enterprise_subscription_inclusion_toggle_with_multiple_orgs(self):
-        """
-        Test that toggling an org's enterprise_subscription_inclusion value on will not turn the course, course run or
-        program on if there is another org that is still off, with that course, course run and program
-        """
-        org = factories.OrganizationFactory(enterprise_subscription_inclusion=False)
-        org2 = factories.OrganizationFactory(enterprise_subscription_inclusion=False)
-        course = factories.CourseFactory(enterprise_subscription_inclusion=False)
-        course_run = factories.CourseRunFactory(
-            course=course,
-            pacing_type='self_paced',
-            enterprise_subscription_inclusion=False
-        )
-        program = factories.ProgramFactory(enterprise_subscription_inclusion=False)
-        program.courses.add(course)
-        program.save()
-
-        course.authoring_organizations.add(org)
-        course.authoring_organizations.add(org2)
-        course.save()
-
-        # Toggle one of the orgs to true
-        org.enterprise_subscription_inclusion = True
-        org.save()
-
-        # Confirm that the course is still False
-        course.refresh_from_db()
-        assert course.enterprise_subscription_inclusion is False
-        assert course_run.enterprise_subscription_inclusion is False
-        assert program.enterprise_subscription_inclusion is False
-
-    def test_org_enterprise_subscription_inclusion_toggle_courserun(self):
-        """Test that toggling an org's enterprise_subscription_inclusion value will toggle the course run"""
-        org = factories.OrganizationFactory(enterprise_subscription_inclusion=True)
-        course_type = CourseType.objects.filter(slug=CourseType.VERIFIED_AUDIT).first()
-        course = factories.CourseFactory(enterprise_subscription_inclusion=True, type=course_type)
-        course_run = factories.CourseRunFactory(
-            course=course,
-            pacing_type='self_paced',
-            enterprise_subscription_inclusion=True
-        )
-
-        course.authoring_organizations.add(org)
-        course.save()
-
-        org.enterprise_subscription_inclusion = False
-        org.save()
-
-        course_run.refresh_from_db()
-        assert course_run.enterprise_subscription_inclusion is False
-
-    def test_org_enterprise_subscription_inclusion_toggle_program(self):
-        """Test that toggling an org's enterprise_subscription_inclusion value will toggle the program"""
-        org = factories.OrganizationFactory(enterprise_subscription_inclusion=True)
-        course_type = CourseType.objects.filter(slug=CourseType.VERIFIED_AUDIT).first()
-        course = factories.CourseFactory(enterprise_subscription_inclusion=True, type=course_type)
-        course.save()
-        course.authoring_organizations.add(org)
-        course.save()
-
-        program_type = ProgramType.objects.get(translations__name_t='XSeries')
-        program = factories.ProgramFactory(enterprise_subscription_inclusion=True, type=program_type)
-        program.courses.add(course)
-        program.save()
-
-        org.enterprise_subscription_inclusion = False
-        org.save()
-
-        program.refresh_from_db()
-        assert program.enterprise_subscription_inclusion is False
 
 
 class OrganizationMappingTests(TestCase):
@@ -4369,3 +4285,68 @@ class RestrictedCourseRunTests(TestCase):
         self.assertEqual(course_run.restricted_run, restricted_course_run)
         self.assertEqual(restricted_course_run.restriction_type, 'custom-b2b-enterprise')
         self.assertEqual(str(restricted_course_run), "course-v1:SC+BreadX+3T2015: <custom-b2b-enterprise>")
+
+
+@ddt.ddt
+class BulkOperationTaskTest(TransactionTestCase):
+
+    def test_bulk_operation_task_creation(self):
+        """
+        Verify that the bulk operation task is created with the correct attributes.
+        """
+        bulk_operation_task = factories.BulkOperationTaskFactory()
+        assert bulk_operation_task.task_type == BulkOperationType.CourseCreate
+        self.assertIsNotNone(bulk_operation_task.uploaded_by)
+        self.assertTrue(bulk_operation_task.csv_file.name.endswith('.csv'))
+
+    def test_task_result_property_with_existing_result(self):
+        """
+        Verify that the task_result method returns the correct TaskResult object
+        """
+        with mock.patch('course_discovery.apps.course_metadata.signals.uuid', return_value='test-task-123'):
+            bulk_operation = factories.BulkOperationTaskFactory()
+            task_result = TaskResult.objects.create(
+                task_id="test-task-123",
+                status="SUCCESS",
+                result='{"message": "Task completed"}'
+            )
+        result = bulk_operation.task_result
+        self.assertEqual(result, task_result)
+
+    def test_task_result_property_with_no_task_result(self):
+        """
+        Verify that the task_result method handles if no task result is found.
+        """
+        bulk_operation = factories.BulkOperationTaskFactory(task_id="non-existent-task")
+        self.assertIsNone(bulk_operation.task_result)
+
+    def test_task_result_property_with_no_task_id(self):
+        """
+        Verify that the task_result method handles if no task ID is associated with the bulk operation.
+        """
+        bulk_operation = factories.BulkOperationTaskFactory(task_id=None)
+        self.assertIsNone(bulk_operation.task_result)
+
+    def test_str_representation(self):
+        """
+        Verify that the string representation of the bulk operation task includes the uploaded_by username.
+        """
+        bulk_operation = factories.BulkOperationTaskFactory()
+        string_output = str(bulk_operation)
+        self.assertIn(bulk_operation.uploaded_by.username, string_output)
+
+    def test_save_assigns_uploaded_by(self):
+        """
+        Verify that the uploaded_by is assigned when saving the bulk operation task.
+        """
+        user = factories.UserFactory()
+        bulk_operation = factories.BulkOperationTaskFactory.build(uploaded_by=None)
+        bulk_operation.save(user=user)
+        self.assertEqual(bulk_operation.uploaded_by, user)
+
+    def test_bulk_operation_task_save_raises_error_without_user_or_uploaded_by(self):
+        """
+        Verify that an error is raised if the user or uploaded_by is not provided when saving.
+        """
+        with pytest.raises(ValueError, match="User is required to save the BulkOperationTask"):
+            factories.BulkOperationTaskFactory(uploaded_by=None)
