+++ course_discovery/apps/course_metadata/tests/test_tasks.py
@@ -6,17 +6,25 @@ from unittest import mock
 import ddt
 import pytest
 from django.conf import settings
-from django.test import TransactionTestCase
-from django_celery_results.models import TaskResult
+from django.db.models.signals import post_save
+from django.test import TestCase, TransactionTestCase
 
+from course_discovery.apps.api.v1.tests.test_views.mixins import OAuth2Mixin
 from course_discovery.apps.course_metadata.choices import BulkOperationStatus, BulkOperationType
+from course_discovery.apps.course_metadata.models import BulkOperationTask, CourseType, Organization, ProgramType
+from course_discovery.apps.course_metadata.signals import (
+    on_bulk_operation_create, update_enterprise_inclusion_for_courses_and_programs
+)
+from course_discovery.apps.course_metadata.tasks import (
+    process_bulk_operation, update_org_program_and_courses_ent_sub_inclusion
+)
 from course_discovery.apps.course_metadata.tests import factories
 
 
 @ddt.ddt
-class BulkOperationTaskTest(TransactionTestCase):
+class ProcessBulkOperationsTest(TransactionTestCase):
     """
-    Unit Test class for testing Celery Tasks and their functionality.
+    Unit Test suite for testing process_bulk_operations celery task.
     """
     def setUp(self):
         factories.PartnerFactory(short_code='test', id=settings.DEFAULT_PARTNER_ID)
@@ -52,6 +60,7 @@ class BulkOperationTaskTest(TransactionTestCase):
             self, loader_class_name, task_type, should_succeed, expected_summary, expected_status,
             mock_course_run_loader, mock_course_loader
     ):
+        post_save.disconnect(on_bulk_operation_create, sender=BulkOperationTask)
         mock_loader = mock_course_loader if loader_class_name == "CourseLoader" else mock_course_run_loader
         if should_succeed:
             mock_loader.return_value.ingest.return_value = expected_summary
@@ -59,6 +68,7 @@ class BulkOperationTaskTest(TransactionTestCase):
             mock_loader.return_value.ingest.side_effect = KeyError("Simulated failure")
 
         bulk_operation_task = factories.BulkOperationTaskFactory(task_type=task_type)
+        process_bulk_operation.apply_async(args=[bulk_operation_task.id])
         bulk_operation_task.refresh_from_db()
 
         mock_loader.assert_called_once()
@@ -66,63 +76,117 @@ class BulkOperationTaskTest(TransactionTestCase):
         self.assertEqual(bulk_operation_task.status, expected_status)
         self.assertEqual(bulk_operation_task.task_summary, expected_summary)
 
-    def test_bulk_operation_task_creation(self):
-        """
-        Verify that the bulk operation task is created with the correct attributes.
-        """
-        bulk_operation_task = factories.BulkOperationTaskFactory()
-        assert bulk_operation_task.task_type == BulkOperationType.CourseCreate
-        self.assertIsNotNone(bulk_operation_task.uploaded_by)
-        self.assertTrue(bulk_operation_task.csv_file.name.endswith('.csv'))
+        post_save.connect(on_bulk_operation_create, sender=BulkOperationTask)
 
-    def test_task_result_property_with_existing_result(self):
-        """
-        Verify that the task_result method returns the correct TaskResult object
-        """
-        with mock.patch('course_discovery.apps.course_metadata.signals.uuid', return_value='test-task-123'):
-            bulk_operation = factories.BulkOperationTaskFactory()
-            task_result = TaskResult.objects.create(
-                task_id="test-task-123",
-                status="SUCCESS",
-                result='{"message": "Task completed"}'
-            )
-        result = bulk_operation.task_result
-        self.assertEqual(result, task_result)
 
-    def test_task_result_property_with_no_task_result(self):
-        """
-        Verify that the task_result method handles if no task result is found.
-        """
-        bulk_operation = factories.BulkOperationTaskFactory(task_id="non-existent-task")
-        self.assertIsNone(bulk_operation.task_result)
+@pytest.mark.django_db
+class EnterpriseSubscriptionInclusionTests(OAuth2Mixin, TestCase):
+    """
+    Test suite for update_org_program_and_courses_ent_sub_inclusion celery task.
+    """
+    def setUp(self):
+        super().setUp()
+        post_save.disconnect(update_enterprise_inclusion_for_courses_and_programs, sender=Organization)
 
-    def test_task_result_property_with_no_task_id(self):
-        """
-        Verify that the task_result method handles if no task ID is associated with the bulk operation.
-        """
-        bulk_operation = factories.BulkOperationTaskFactory(task_id=None)
-        self.assertIsNone(bulk_operation.task_result)
+    def tearDown(self):
+        post_save.connect(update_enterprise_inclusion_for_courses_and_programs, Organization)
+        super().tearDown()
 
-    def test_str_representation(self):
-        """
-        Verify that the string representation of the bulk operation task includes the uploaded_by username.
-        """
-        bulk_operation = factories.BulkOperationTaskFactory()
-        string_output = str(bulk_operation)
-        self.assertIn(bulk_operation.uploaded_by.username, string_output)
+    def test_org_enterprise_subscription_inclusion_toggle_course(self):
+        """Test that toggling an org's enterprise_subscription_inclusion value will turn courses in the org on"""
+        org = factories.OrganizationFactory(enterprise_subscription_inclusion=True)
+        course_type = factories.CourseTypeFactory(slug=CourseType.VERIFIED_AUDIT)
+        course = factories.CourseFactory(enterprise_subscription_inclusion=True, type=course_type)
+        course.authoring_organizations.add(org)
+        course.save()
 
-    def test_save_assigns_uploaded_by(self):
-        """
-        Verify that the uploaded_by is assigned when saving the bulk operation task.
-        """
-        user = factories.UserFactory()
-        bulk_operation = factories.BulkOperationTaskFactory.build(uploaded_by=None)
-        bulk_operation.save(user=user)
-        self.assertEqual(bulk_operation.uploaded_by, user)
+        org.enterprise_subscription_inclusion = False
+        org.save()
 
-    def test_bulk_operation_task_save_raises_error_without_user_or_uploaded_by(self):
+        update_org_program_and_courses_ent_sub_inclusion.apply_async(
+            args=[org.id, org.enterprise_subscription_inclusion]
+        )
+
+        course.refresh_from_db()
+        assert course.enterprise_subscription_inclusion is False
+
+    def test_org_enterprise_subscription_inclusion_toggle_with_multiple_orgs(self):
         """
-        Verify that an error is raised if the user or uploaded_by is not provided when saving.
+        Test that toggling an org's enterprise_subscription_inclusion value on will not turn the course, course run or
+        program on if there is another org that is still off, with that course, course run and program
         """
-        with pytest.raises(ValueError, match="User is required to save the BulkOperationTask"):
-            factories.BulkOperationTaskFactory(uploaded_by=None)
+        org = factories.OrganizationFactory(enterprise_subscription_inclusion=False)
+        org2 = factories.OrganizationFactory(enterprise_subscription_inclusion=False)
+        course = factories.CourseFactory(enterprise_subscription_inclusion=False)
+        course_run = factories.CourseRunFactory(
+            course=course,
+            pacing_type='self_paced',
+            enterprise_subscription_inclusion=False
+        )
+        program = factories.ProgramFactory(enterprise_subscription_inclusion=False)
+        program.courses.add(course)
+        program.save()
+
+        course.authoring_organizations.add(org)
+        course.authoring_organizations.add(org2)
+        course.save()
+
+        # Toggle one of the orgs to true
+        org.enterprise_subscription_inclusion = True
+        org.save()
+
+        update_org_program_and_courses_ent_sub_inclusion.apply_async(
+            args=[org.id, org.enterprise_subscription_inclusion])
+
+        # Confirm that the course is still False
+        course.refresh_from_db()
+        assert course.enterprise_subscription_inclusion is False
+        assert course_run.enterprise_subscription_inclusion is False
+        assert program.enterprise_subscription_inclusion is False
+
+    def test_org_enterprise_subscription_inclusion_toggle_courserun(self):
+        """Test that toggling an org's enterprise_subscription_inclusion value will toggle the course run"""
+        org = factories.OrganizationFactory(enterprise_subscription_inclusion=True)
+        course_type = factories.CourseTypeFactory(slug=CourseType.VERIFIED_AUDIT)
+        course = factories.CourseFactory(enterprise_subscription_inclusion=True, type=course_type)
+        course_run = factories.CourseRunFactory(
+            course=course,
+            pacing_type='self_paced',
+            enterprise_subscription_inclusion=True
+        )
+
+        course.authoring_organizations.add(org)
+        course.save()
+
+        org.enterprise_subscription_inclusion = False
+        org.save()
+
+        update_org_program_and_courses_ent_sub_inclusion.apply_async(
+            args=[org.id, org.enterprise_subscription_inclusion])
+
+        course_run.refresh_from_db()
+        assert course_run.enterprise_subscription_inclusion is False
+
+    def test_org_enterprise_subscription_inclusion_toggle_program(self):
+        """Test that toggling an org's enterprise_subscription_inclusion value will toggle the program"""
+        post_save.disconnect(update_enterprise_inclusion_for_courses_and_programs, sender=Organization)
+
+        org = factories.OrganizationFactory(enterprise_subscription_inclusion=True)
+        course_type = factories.CourseTypeFactory(slug=CourseType.VERIFIED_AUDIT)
+        course = factories.CourseFactory(enterprise_subscription_inclusion=True, type=course_type)
+        course.authoring_organizations.add(org)
+        course.save()
+
+        program_type = factories.ProgramTypeFactory(slug=ProgramType.XSERIES)
+        program = factories.ProgramFactory(enterprise_subscription_inclusion=True, type=program_type)
+        program.courses.add(course)
+        program.save()
+
+        org.enterprise_subscription_inclusion = False
+        org.save()
+
+        update_org_program_and_courses_ent_sub_inclusion.apply_async(
+            args=[org.id, org.enterprise_subscription_inclusion])
+
+        program.refresh_from_db()
+        assert program.enterprise_subscription_inclusion is False
